+++ internal/ceres/schur_jacobi_preconditioner.h
@@ -89,7 +89,7 @@ class SchurJacobiPreconditioner : public BlockSparseMatrixPreconditioner {
 
   // Preconditioner interface.
   void RightMultiply(const double* x, double* y) const final;
-  int num_rows() const final;
+  int64_t num_rows() const final;
 
  private:
   void InitEliminator(const CompressedRowBlockStructure& bs);
